pyproject.toml
[tool.poetry]
name = "adhd-productivity-assistant"
version = "0.1.0"
description = "An ADHD Productivity Assistant with Quart backend and TypeScript frontend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
quart = "0.18.4"
werkzeug = "2.3.7"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"
tiktoken = "^0.5.2"
hypercorn = "^0.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
